pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glean"
version = "0.2.0"
description = "Ingest media, extract knowledge to Obsidian vault, search with natural language."
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "python-dotenv>=1.0",
    "google-genai>=1.0",
    "ffmpeg-python>=0.2",
    "imageio-ffmpeg>=0.5",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
glean = "glean.cli:cli"

[tool.setuptools.packages.find]
include = ["glean*"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests that download videos or do heavy I/O (deselect with '-m \"not slow\"')",
]
